Refactor into one combined stage
